moved bathroom image to cardview
